archive/requirements.txt
# Core Google AI/ADK
google-adk>=1.16.0
google-cloud-aiplatform>=1.121.0
google-genai>=1.0.0
google-generativeai>=0.8.0

# Google Auth & APIs
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-cloud-storage>=2.10.0

# Core utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.0.0
pyyaml>=6.0
structlog>=23.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# ========================================
# Phase 2+ Dependencies (Uncomment when needed)
# ========================================

# Web Framework (Phase 3+)
# fastapi>=0.100.0
# uvicorn[standard]>=0.20.0
# python-multipart>=0.0.6
# jinja2>=3.1.0

# Data Processing (Phase 2+)
# pandas>=2.0.0
# numpy>=1.24.0

# Image/Video Processing (Phase 2+)
# pillow>=10.0.0
# opencv-python>=4.8.0

# Database (if needed later)
# sqlalchemy>=2.0.0
# alembic>=1.10.0

